requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
google-cloud-vertexai==0.0.1
google-cloud-aiplatform==1.36.4
google-cloud-speech==2.24.0
pytz==2024.1
